RAG Application/requirements.txt
streamlit
PyPDF2
langchain
langchain-community
langchain-google-genai
faiss-cpu
python-dotenv
google-generativeai